Fix caracteristicas de la app y borrado de funciones o comentarios sin uso

+++ Proyectofinalbebedero/app/src/main/java/com/example/proyecto_final_bebedero/FirebaseDatabaseInterface.kt
@@ -9,7 +9,7 @@ interface FirebaseDatabaseInterface {
 
     fun getWaterLevelsSortedByDate(listener: (Map<String, List<Double>>) -> Unit) //listener lambda function
 
-    fun getMinWaterLevel(listener: (Double) -> Unit) //listener lambda function
+    fun getMinWaterLevel(listener: (Int) -> Unit) //listener lambda function
 
     fun getWaterLevels(listener: (List<Double>) -> Unit) //listener lambda function
